% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_assets.R
\docType{data}
\name{asset_choices}
\alias{asset_choices}
\title{Central Asset Choices for UI Dropdowns}
\format{
An object of class \code{list} of length 8.
}
\usage{
asset_choices
}
\description{
Central Asset Choices for UI Dropdowns
}
\keyword{datasets}
