% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataapp-package.R
\docType{package}
\name{dataapp-package}
\alias{dataapp}
\alias{dataapp-package}
\title{dataapp: An Amazing Shiny App}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: firstname lastname \email{your@email.com}

}
\keyword{internal}
