% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset1.R
\docType{data}
\name{dataset1}
\alias{dataset1}
\title{Random_LabValuesInfo_2020.tsv.}
\format{
Sample dataset 1
\describe{

}
}
\source{
\url{https://github.com/fontes-lab/RShinyProgrammingExercise}
}
\usage{
dataset1
}
\description{
A dataset containing sample medical data.
}
\keyword{datasets}
