% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset2.R
\docType{data}
\name{dataset2}
\alias{dataset2}
\title{Random_PatientLevelInfo_2020.tsv.}
\format{
Sample dataset 2
\describe{

}
}
\source{
\url{https://github.com/fontes-lab/RShinyProgrammingExercise/}
}
\usage{
dataset2
}
\description{
A dataset containing sample medical data.
}
\keyword{datasets}
