% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_helpers.R
\name{make_dropdown}
\alias{make_dropdown}
\title{Helper function for preparing dropdown menus}
\usage{
make_dropdown(drop_id, drop_label, drop_choices)
}
\arguments{
\item{drop_id, drop_label, drop_choices}{selectize input elements}
}
\description{
Helper function for preparing dropdown menus
}
