% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_backtester.R
\name{mod_backtester_server}
\alias{mod_backtester_server}
\title{Backtester Server}
\usage{
mod_backtester_server(id, opt_payload = NULL)
}
\arguments{
\item{id}{NA}

\item{opt_payload}{Reactive payload imported from mod_optimizer securely linking module states}
}
\description{
Core execution server for parsing inputs and communicating with Rcpp trading engine.
}
