% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_backtester.R
\name{mod_backtester_ui}
\alias{mod_backtester_ui}
\title{Backtester UI}
\usage{
mod_backtester_ui(id)
}
\arguments{
\item{id}{NA}
}
\description{
UI module for configuring and running the C++ optimized backtester
}
