% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_data_explorer.R
\name{mod_data_explorer_server}
\alias{mod_data_explorer_server}
\title{Data Explorer Server}
\usage{
mod_data_explorer_server(id)
}
\arguments{
\item{id}{NA}
}
\description{
Server logic for retrieving, formatting, and charting market data
}
