% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_optimizer.R
\name{mod_optimizer_server}
\alias{mod_optimizer_server}
\title{Screener Server}
\usage{
mod_optimizer_server(id, screen_payload = NULL)
}
\arguments{
\item{id}{NA}

\item{screen_payload}{Input payload reacting dynamically via market screen pipelines.}
}
\description{
Server logic for evaluating screener strategies across various targets and base assets
}
