% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_screener.R
\name{mod_screener_server}
\alias{mod_screener_server}
\title{Screener Server}
\usage{
mod_screener_server(id)
}
\arguments{
\item{id}{NA}
}
\description{
Server logic for evaluating screener strategies across various targets and base assets
}
