% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_screener.R
\name{mod_screener_ui}
\alias{mod_screener_ui}
\title{Screener UI}
\usage{
mod_screener_ui(id)
}
\arguments{
\item{id}{NA}
}
\description{
UI for the market screener module allowing for testing various strategies across N-dimensions
}
