% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_strategy_discovery.R
\name{mod_strategy_discovery_server}
\alias{mod_strategy_discovery_server}
\title{Strategy Discovery Server}
\usage{
mod_strategy_discovery_server(id)
}
\arguments{
\item{id}{NA}
}
\description{
Strategy Discovery Server
}
