% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{server}
\alias{server}
\title{Server function}
\usage{
server(input, output, session)
}
\arguments{
\item{input}{input for app}

\item{output}{output for app}

\item{session}{session info}
}
\description{
code to execute server scripts
}
