\name{testapp-package}
\alias{testapp-package}
\alias{testapp}
\docType{package}
\title{TestApp}
\description{}
\details{
\tabular{ll}{
Package: \tab testapp\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2014-07-11\cr
License: \tab What license is it under?\cr
}
}
\author{Jan Fait}
\references{}
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{}
