% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_app}
\alias{run_app}
\title{Run the app}
\usage{
run_app(path = ".")
}
\arguments{
\item{path}{Path to interact with.}
}
\value{
Called for its side effect.
}
\description{
Run the app
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
run_app("/")
\dontshow{\}) # examplesIf}
}
