% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runn_app.R
\name{run_app}
\alias{run_app}
\title{Title}
\usage{
run_app(path = ".")
}
\description{
Title
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
run_app()
\dontshow{\}) # examplesIf}
}
